+++ packages/app/components/settings/tabs/payouts.tsx
@@ -1,9 +1,7 @@
-import { Platform, Linking } from "react-native";
+import { Platform } from "react-native";
 
 import { Button } from "@showtime-xyz/universal.button";
-import { LinkOut } from "@showtime-xyz/universal.icon";
 import { Image } from "@showtime-xyz/universal.image";
-import { Pressable } from "@showtime-xyz/universal.pressable";
 import { useRouter } from "@showtime-xyz/universal.router";
 import Spinner from "@showtime-xyz/universal.spinner";
 import { TabScrollView } from "@showtime-xyz/universal.tab-view";
@@ -11,7 +9,7 @@ import { Text } from "@showtime-xyz/universal.text";
 import { View } from "@showtime-xyz/universal.view";
 
 import { useOnboardingStatus } from "app/components/payouts/hooks/use-onboarding-status";
-import { useStripeAccountLink } from "app/components/payouts/hooks/use-stripe-account-link";
+import { PayoutSettings } from "app/components/payouts/payout-settings";
 import { payoutsRedirectOrigin } from "app/components/payouts/payouts-setup";
 
 import { SettingsTitle } from "../settings-title";
@@ -155,41 +153,5 @@ export const Payouts = ({ index = 0 }: { index: number }) => {
         </View>
       </View>
     </SettingScrollComponent>
-  );
-};
-
-export const PayoutSettings = ({
-  refreshUrl,
-  returnUrl,
-}: {
-  returnUrl: string;
-  refreshUrl: string;
-}) => {
-  const stripeAccountLink = useStripeAccountLink();
-
-  return (
-    <Pressable
-      onPress={async () => {
-        const value = await stripeAccountLink.trigger({
-          refresh_url: refreshUrl,
-          return_url: returnUrl,
-        });
-        if (Platform.OS === "web") {
-          window.location.href = value.url;
-        } else {
-          Linking.openURL(value.url);
-        }
-      }}
-    >
-      <View tw="flex-row items-center" style={{ columnGap: 4 }}>
-        <Image
-          source={require("app/components/payouts/stripe-logo.png")}
-          height={20}
-          width={20}
-        />
-        <Text tw="font-semibold text-[#6672e4]">View payout settings</Text>
-        <LinkOut height={16} width={16} color="#6672e4" />
-      </View>
-    </Pressable>
   );
 };
